// src/types/Coin.d.ts
export type Graph = {
  date: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
};

export type PriceDataPoint = {
  timestamp: number;
  price: number;
};